refactor(calendar): migrate local composables
